+++ src/mlxtranOp.c
@@ -183,13 +183,7 @@ void trans_mlxtran_op(const char* parse){
   errP = curP;
   eBufLast = 0;
   gBufFree=0;
-  {
-    size_t _gBuf_len = strlen(gBuf);
-    if (_gBuf_len > (size_t)INT_MAX) {
-      Rf_error(_("input too large to parse (exceeds INT_MAX bytes)"));
-    }
-    _pn= dparse(curP, gBuf, (int)_gBuf_len);
-  }
+  _pn= dparse(curP, gBuf, monolix2rxParseLen(gBuf, "mlxtran option"));
   if (!_pn || curP->syntax_errors) {
   } else {
     wprint_parsetree_mlxtran_op(parser_tables_mlxtranOp , _pn, 0, wprint_node_mlxtran_op, NULL);